requirements.txt
pillow
pytesseract
sentence-transformers
faiss-cpu
pdf2image
PyMuPDF
tiktoken
google-genai
streamlit
